+++ sdk/trade360-java-sdk/src/main/java/eu/lsports/trade360_java_sdk/common/entities/fixtures/Participant.java
@@ -36,4 +36,16 @@ public class Participant {
      */
     public int isActive = -1;
 
+    /**
+     * The participant shirt color details.
+     * This can be {@code null} if not set.
+     */
+    @Nullable public UniformColor shirtColor;
+
+    /**
+     * The participant goalkeeper shirt color details.
+     * This can be {@code null} if not set.
+     */
+    @Nullable public UniformColor goalKeeperShirtColor;
+
 }