+++ sdk/trade360-java-sdk/src/test/java/eu/lsports/trade360_java_sdk/common/entities/fixtures/ParticipantTest.java
@@ -12,12 +12,24 @@ class ParticipantTest {
         participant.position = "Forward";
         participant.rotationId = 7;
         participant.isActive = 1;
+        participant.shirtColor = new UniformColor();
+        participant.shirtColor.primary = "ffffff";
+        participant.shirtColor.number = "000000";
+        participant.shirtColor.outline = "ffffff";
+        participant.goalKeeperShirtColor = new UniformColor();
+        participant.goalKeeperShirtColor.primary = "ffff00";
+        participant.goalKeeperShirtColor.number = "000000";
+        participant.goalKeeperShirtColor.outline = "ffff00";
 
         assertEquals(42, participant.id);
         assertEquals("John Doe", participant.name);
         assertEquals("Forward", participant.position);
         assertEquals(7, participant.rotationId);
         assertEquals(1, participant.isActive);
+        assertNotNull(participant.shirtColor);
+        assertEquals("ffffff", participant.shirtColor.primary);
+        assertNotNull(participant.goalKeeperShirtColor);
+        assertEquals("ffff00", participant.goalKeeperShirtColor.primary);
         assertNotNull(participant);
     }
 
@@ -34,6 +46,8 @@ class ParticipantTest {
         assertNull(participant.name);
         assertNull(participant.position);
         assertNull(participant.rotationId);
+        assertNull(participant.shirtColor);
+        assertNull(participant.goalKeeperShirtColor);
     }
 
     @Test